Move SyntheticShapeNaming logic into LoaderUtils
